common: reduce allocations in ECLogger (1/2)
When ECLogger_{File,Tee}::LogVA has finished building its message
buffer, this is passed to Log(unsigned int, const std::string &),
thereby upgrading to a std::string. This allocation is unnecessary.

Add the new functions "log", "logf" and "logv" (in the style of
printf, vprintf) that work exclusively with C strings. Wrappers
for the old names are provided.