daemons: drop polling for POLLRDHUP
The use of POLLRDHUP is uncalled for. In Linux's implementation of
TCP and PF_LOCAL sockets, POLLIN is always set when POLLRDHUP is. KC
always tests for POLLIN at the same time it tests for POLLRDHUP. In
addition, when KC was still using "select", there was not even a
POLLRDHUP equivalent. So drop it altogether.