dbadm: reverse-run through defrag list
The defrag operation as it currently stands can require up to 31484
moves, for example if IDs {2..31485} are used. It would iterate
through them in ascending order and "downshift" each one in turn.

This keeps the order between names, which is absolutely unnecessary.
By iterating in descending order, low free spots will be populated
and names that (then) do not need a move anymore will stay where they
are. The upper bound is now 15742 moves (IDs {15743..31485} used).