libicalmapi: ensure organizer get resolved to ZARAFA user if local on reply
ICAL_METHOD_REPLY is the only kind of method that is not handled by
HrAddRecipients. We desire if the user is a local user, it is represented
by the ZARAFA address type. However, the current implementation only allowed
SMTP, but since we got the resolve_organizer function we will take advantage
of that, to get the ZARAFA address type if present.