libicalmapi: get attendance status from PROP_RESPONSESTATUS
When WebApp accepts a meeting invitation it stores the attendance status
(accept, tentative, etc) in PROP_RESPONSESTATUS, only the responses of the
meeting requests is stored in PR_RECIPIENT_TRACKSTATUS if you are the
organizer. If a user accepts a meeting using WebApp this is not visible
in Caldav without this change.