notifications: add object-type filter flags
The current notification system seems limited in that it's not
possible to get notifications for message changes within a
specific folder, or to only get notifications for folder changes
within a store (so excluding message and store changes, for
example).

It is actually somehow possible using IECChangeAdvisor, but this
is not ideal as it results in different kinds of notifications,
and for example entryids are not available for deletion events
(as is a known issue with ICS).

To remedy this, we add two custom flags, to override the default
object types for which notifications are generated. So now, when
generating notifications for a specific folder, we can specify
if we want message changes relating to this folder as well, or
only message or only folder changes. Also, for store-level
notifications, it is now possible to only enable folder or message
changes.

When neither of the two flags is set, behaviour should be the
same as before.

References: KC-1150