dbadm: reorder defrag queries for safer hard termination
Hard-terminating a defrag is now safe(r).

Before: ID change in "names" first, renames in "properties", renames
in other tables. If the program is interrupted between step 2 and 3,
rows in "tproperties" with @oldtag are considered lost.

After: Renames in "properties", renames in all other tables, then ID
change in "names". If the program is interrupted between step 2 and
3, @oldtag is still considered in use while @newtag is not. A rerun
of np-defrag will therefore consider renaming more @newtag rows to
@oldtag.

References: KC-1155