libserver: implement object caching by using the cell cache
The loadObject operation is quite heavy. If the SQL log is closely monitored it
is quite clear that the same queries are issued over and over again. Since all
operations (read, write, delete) go through the server, it is possible to have
an up-to-date cache of objects in the server. This is what has been implemented
in this commit.