common: force Unicode for internal string translations
Because gettext is tailored for display purposes, it defaults to
LC_CTYPE for the encoding (only the language is taken from
LC_MESSAGES). That does not work well with Kopano, since e.g.
libkcclient uses gettext for behind-the-scenes translations, so a
Unicode-capable charset like UTF-8 should be set as the default
gettext output encoding.

References: KC-1140