ical: be more intelligent about parsing Content-Type
kopano-ical was unable to parse a Content-Type header that used a
legit charset enclosed in dquotes, "text/html; charset=\"utf-8\"".

Use the same logic from libkcinetmapi. Not wanting to change the ABI
tag by adding a new function to libkcutil, the function is replicated
and later deduplicated in master.

References: KC-1198