dagent: evaluate PR_EMAIL_ADDRESS for forwards
The use of PR_SMTP_ADDRESS was introduced in SVN r25138 for ZCP-2296,
without any comment on why that property instead of the more
prominent (and documented for HrGetAddress) PR_EMAIL_ADDRESS.

ZCSAB addresses worked in 8.1 because the forwarding test did not
interfere. The zcontacts provider seems not needed for resolution, as
the ZCSAB provider seems to only ever store SMTP type addresses,
never ZARAFA ones.

Fixes: bp/8.2.0-315-gd1259edb
References: KC-1037