spooler, dagent: change log_raw_message_path to /var/lib/kopano
When systemd.unit's PrivateTmp is in use, the temporary directory for the service
will be something like
"/tmp/systemd-private-dde339fa502345f59e2b4b648799823d-kopano-server.service-18OLMJ",
and be bind-mounted to /tmp for the process. Similarly for /var/tmp.

For .eml files, I would argue that they should outlive a service
restart, or more importantly, a process termination due to signal,
and as such, the file dumps should be placed outside of the temporary
directory.