client: avoid strict aliasing violation in HrGetServerURLFromStoreEntryId
The expression

	peid_V0->szServer

leads to undefined behavior, and likewise would

	({ EID_V0 e; memcpy(&e, lpEntryId, sizeof(e)); e.szServer; })

because of the shortness of szServer *in EID/EID_V0*. A direct
memcpy/assign from lpEntryId is needed.

The issue has been in KC all along, though only seems to manifest
with g++ 8.x so far. (Assignment of rServerTemp was optimized out.)

References: KC-930