libserver: fix rebase fallout (crash) in ECGenericObjectTable::FindRow
d4125 was written first but not merged. Then, -Wshadow warnings got
fixed up, and the local lpsRestrict variable was renamed rt, at which
point lpsRestrict referred to the class member. Then d4125 was
applied, and it compiled without a diagnostic because of the prior
wanton variable shadowing.

Fixes: kopanocore-8.6.80-1249-gd41256b1b
References: KC-930, KF-1635