misc: clear out dead code
CID-187443 (#1 of 1): Identical code for different branches
(IDENTICAL_BRANCHES)identical_branches: The same code is executed
when the condition hr != 0 is true or false, because the code in the
if-then branch and after the if statement is identical. Should the if
statement be removed?
CID-187444 (#1 of 1): Identical code for different branches
(IDENTICAL_BRANCHES)