libserver: re-work ALTER statements for utf8mb4 conversion
Unspecific reports suggested that there might be some past MySQL
version which is unable to see that an ALTER TABLE CONVERT TO statment
on a table without any CHAR-based table fields is a conversion no-op.

Those tables only need an ALTER TABLE DEFAULT CHARSET, so just do
that.

References: KC-725