libserver: improve cache/db consistency for copyFolder
For copyFolder, avoid the cache being temporarily inconsistent with the
database between transaction commit and cache invalidation.

This is a problem for example when making use of the changes table, where
we sync a change, but as the cache may be inconsistent with the database,
we don't always get the latest state (in case of MSR), resulting in subtle
issues, potentially leading to data loss..

TODO Find general solution to include cache invalidation in wider transaction?