ECtools: convert EX addresses to SMTP addresses
EX addresses don't work in the WebApp and contain legacyExchangeDN in
it's email address. Certain properties have to be rewritten to be useful
in the WebApp and other clients namely, the sender, sent representing,
received, received reprenstenting and the whole recipient table.

The original email address and display name can be found in different
properties in the PST. The PR_SEARCH_KEY, PR_ENTRYID has to be
re-created since it contains the legacyExchangeDN and is no longer
valid.