migration-pst: add --create-mapping, --mapping
Add the option to generate a mapping file from multiple PST's with
--create-mapping which creates a JSON file with key legacyExchangeDN and
value the found email address from the recipient table.

With --mappping this mapping can be used to find an email address for an
item which has no email address. If no email address is found an log
message will be emitted.