dbadm: add routine for shrinking indexed columns
Table indexes have varying limits, depending on the MySQL-InnoDB
on-disk format that they are in. This seems to prevent the schema
upgrade to v69 for a user (KF-1934). Unicode SMP support (KC-725) ran
into more or less the same issue, so here is a shrinker function.

References: KC-725, KC-1357, KF-1934