libserver: do not mark object as complete when it's not
The purpose of the loadObject RPC is general hierarchy discovery. The
server sends gratitious propvals along with it to reduce the number
of roundtrips. loadObject has a built-in cutoff limit (MAX_PROP_SIZE:
32K). LoadObject/PrepareReadProps returns the property list and, for
selected properties, their values.

If client code requests any propvals not received by the above
mechanism (i.e. "large" props), further propvals are loaded using the
loadProp RPC as originally designed.

Commit c0b1f0794 added cache-on-read; before, there was only
cache-on-write. The PrepareReadProps propvals were also entered into
the server-side cache, and then the cell was marked as "complete",
even though it was not complete.

This patch adds a check for this condition.

Fixes: kopanocore-8.6.80-869-gc0b1f0794
References: KC-1359, KF-1940