rest: basic pluggable backends
make it so that ldap/imap/caldav can be used
instead of kopano, or any combination of these.

examples of routing that needs to happen:

/users -> pick 'directory' backend (ldap/kopano)
/users/userid/calendar -> pick default 'calendar' backend
/me/messages -> pick default 'mail' backend

/users/imap.frits/messages -> pick 'imap' backend
/users/caldav.frits/calendar -> pick 'caldav' backend

there is some complication esp. with the user resource,
because I don't know how to split it at the routing level
(user/messages goes to 'mail', user/calendar to 'calendar').