provider: change ABEID to use a Flexible Array Member
Starting with version 8, gcc transforms strlen calls on 1-sized
arrays that are not the last member to yield zero.

	struct A { char szExId[1], szPadding[3]; };
	int len(struct A *a) { return strlen(a->szExId); }

In practice, this causes AB ENTRYIDs to be truncated in e.g.
UnWrapServerClientABEntry, to the point that users become referenced
by their userid alone, and that breaks in multiserver environment.

The struct is wrongly implementing the flexible array member scheme
(C11 §6.7.2.1). A::szExId must be the *last* element.

Padding is hereby externalized by rewriting CbNewABEID to add it,
rather than to subtract it. ABEID becomes a struct that is not
instantiatable, and only usable through a pointer. A clone,
ABEID_PAD, is created to support the few fixed-size ABEIDs.

Fixes: SVN r2956
References: KS-42325, KC-1386