libserver: select KCOID SSO user via exID
SSO logon with KCOID receives ABEID values from the access token to
identify the desired user. To ensure functionality with multi server
Kopano setups, this change updates the selector from the local ID to the
external ID part of the provided ABEID value as it is not guaranteed nor
desired that the access token value must be generated on the same
server. Lookup is from cache to avoid an additional SQL query.

Reference: KC-1384