client: fix mismatching types
Unspecified UCS system showed this trace for 8.6.8:

 f4. /usr/lib/kopano/libkcclient.so(+0x86e10) [0x7f805986ae10]
 f5. /usr/lib/kopano/libkcclient.so(+0x7bc03) [0x7f805985fc03]
 f6. /usr/sbin/kopano-gateway(+0x2ccc3) [0x560d1a9bccc3]

Assuming it was UCS 4.3, this points to ECMAPIProp::DefaultMAPIGetProp. Looking
around, I find that, for example, ECABContainer::ECABContainer stores a
pointer-to-ECABContainer (HrAddPropHandlers does not adjust it in any way):

	HrAddPropHandlers(PR_AB_PROVIDER_ID, DefaultABContainerGetProp, DefaultSetPropComputed, this);

while at the same time, ECABContainer::DefaultABContainerGetProp reads it as a
ECABProp instead.

	auto lpProp = static_cast<ECABProp *>(lpParam);

That reeks of undefined behavior and only works by accident (if it does). To
resolve this, HrAddPropHandlers and all related functions will take a
pointer-to-ECGenericProp instead of pointer-to-void.

References: KF-2151