inetmapi: revert vmime::word wrapping
25843's commit message said """[wrap it] in vmime::word, so you don't
get split up normal/qp strings""". A split normal/QP string likely
refers to a vmime::text with multiple words in different encodings.

Such a case with multiple words in a text objects should not be
occurring, because the text object as returned from
getVmimeTextFromWide is never appended to, AFAICS.

Fixes: SVN r25843
References: KC-1430