migration-pst: skip over corrupt entryids
For a customer PST, there is a property on the subtree folder,
PidTagIpmWastebasketEntryId, which contains (AFAICS) a corrupt
entryid, causing the entire folder (the entire subtree in this
case) to be skipped when importing.

Since we can easily continue in such cases, let's just skip such
properties, log an error and continue.