inetmapi: use PR_RECORD_KEY for Message-ID generation
As per RFC 5322 page 25 § 3.6.4 ¶5, the msg-id is supposed to change
for different versions of the same message. As such,
PR_LAST_MODIFICATION_TIME henceforth factors into the construction
the Message-Id in some capacity.

Second, PR_SEARCH_KEY is a bad pick for message IDs. It does not
change on CopyMessage(), and furthermore, such copies can be edited
without incurring a modification PR_SEARCH_KEY. MSDN also specifies
that the search key essentially consists of some email address, which
makes it anything but unique.

References: KC-1458