pyko: handle broken recurrence objects
Some recurrence blobs can be broken and python-kopano would just throw
an exception and thereby for example breaking a JSON request in grapi.
Handle at least unpack errors by catching it in _parse and logging it,
also add a condition to check if the blob was succesfully parsed which
can be used to determine if we have recurrences.