dagent: set Bcc/RecipMe flags appropriately
In Internet Mail, the Bcc field is a MUA UI-specific field that gets
transformed to a multi-recipient envelope when contacting an SMTP
server or a local maildrop program. In MAPI, the Bcc field is also a
sending-side field, and kopano-spooler equally expands this to a
multi-recipient SMTP envelope.

There is never a Bcc: field in the SMTP DATA part. Checking for a
"Bcc" field in VMIMEToMAPI is pointless and henceforth removed. For
the same reason, the recipient tables of incoming messages that were
just transformed from IM to MAPI never have a MAPI_BCC recipient.
Therefore, having dagent check for a MAPI_BCC entry is pointless.

Instead, Bcc must be determined by the absence of one's address from
all three of To, Cc, and distribution lists.

References: KC-319, KW-840