msr: handle invalid commands send by client
The custom protocol is quite simple but lacks proper error handling in
the msr daemon. Handle an empty command which leads to an empty list and
add error handling for the amount of arguments for the
ADD/DETAILS/REMOVE. In theory these errors should not occur since the
client is controlled by us but it is nice to handle.