pyko: remove UTC meetingrequest base date conversion
A events base date needs to be in the timezone of the event itself
and thus cannot be converted to UTC since that might be the day
before or after based on the events timezone. This change removes
invalid UTC conversion here so that it is ensured that the date
does not change.