storeadm: try loading UTF-8 locale if plain locale fails
storeadm does not care about the particular encoding of the
LC_MESSAGES locale, because encoding is taken from LC_CTYPE anyway.
But the setlocale() is exact and will fail when "<locale>.<encoding>"
exists but "<locale>" does not.

Adjust storeadm to re-try the locale string with ".UTF-8" appended in
the hope the locale setting succeeds. Not bothering with other
encodings like ".euckr" at this time.

References: KC-1429