build: cure g++-8 build failure with xapian 1.2 API
g++-8 from devtoolset has a beef with using the xapian 1.2.22 API the xapian way:

idx_xapian.cpp: In member function 'virtual void KC::ECXapianIndexer::commit(const string&)':
idx_xapian.cpp:145:29: error: invalid conversion from 'int' to 'Xapian::TermGenerator::flags' [-fpermissive]
  tg.set_flags(!sugg.empty() ? tg.FLAG_SPELLING : 0);

The ?:0 construct appears in the source code of Xapian as well, but
of course that is precompiled (with g++-4.8.5 or so). No further
investigation is made, and the short-term measure is to just rewrite
the ?: as a full if statement.