pyko: ensure correct time of recurring events
Recurring events happen in a certain time zone which is stored
together with the event as a property. Whenever pyko reads or
writes recurring event timezone dependent properties, those data
values need conversion to local time which is what other pyko
code expect. This change adds the conversion where it was missing
and thus ensures that recurring events use the correct time
independent of the system's time zone.