php-ext: add mapi_vcftomapi2 function to import multiple vCards
Derived from commit kopanocore-8.7.81-165-gd1ff4bbf9 .

A method signature change was needed to support adding multiple vcards
from one file. This was accomplished by creating a new method.

mapi_vcfstomapi takes a session, a message store, a folder and
the vcf text. It will parse the vcf text and return an array of
messages that are created inside the specified folder or FALSE
on error.

References: KC-1487