pyko: correctly close threadsafe logger queue on shutdown
When shutting down in a multi threaded or multi process environment,
the shutdown order is important. Sometimes the shutdown triggered
a race condition where some thread or process logs when the logger
in another thread/process was already closed. This change ensures
the correct closing/join order to avoid the issue.