build: print versionheader errors to stderr
The versionheader.sh script as a bunch of checks which can result
in error messages. Previously the script would happliy write those
to stdout and thus writing them to the c header file it creates.
This change echos error messages on stderr so they become visible
while at the same time avoids generation of an invalid header file.