pyko: use pytz for timezones consistently
Pytz timezones with DST can only be used properly with the
localize method. This change removes all use of dateutil.tz and
consistently uses pytz for time zone conversion.

To figure out the local time zone, a new dependency to tzlocal
cannot be avoided as the previous behavior was not able to
compute local time properly in a usable fashion for conversion of
arbitrary datetimes and thus had to go away.