pyko: handle permission error for users in multi tenant more
In multi tenant mode the gab is local to a company. As administrator
access is required to get the individual companies gab, this change
ignores the error when not having administrator rights and then falls
back to leave the reducing of results to the gab instead based on
the authentication session. This works fine when the company requested
is matching the one of the authenticated user.

The change here can only be considered a workaround. A better
solution would be to only even do try to select a different company
gab when the company in question is actually a different one from
the authentication session user's company.

Fixes: KC-1575