client: avoid password logon after KCOIDC logon has failed
KCOIDC logon is selected by the client when a flag is passed. If that
flag is passed, it means that the password value is not the password,
but instead the OIDC access token. If logon with that token fails, the
whole logon has failed and it is pointless to try to logon using the
OIDC token as a password.

With this change, client logon function immediately exits with
an error after KCOIDCLogon has failed.