libserver: avoid generating K-1504 warnings
The presence of ldap_groupmembers_relation_attribute = uid in
ldap.openldap.cfg is a "promise" to ldapplugin that all members of a
group have this attribute present. Conversely, an LDAP object that
does not have the attribute can never be a member of a group, as the
group literally has no way to refer to the member using that
attribute.

So when ECSecurity attempts to a group-in-group lookup, this will
fail, and emit lots of K-1504 warnings because a (pure) RFC2307-style
group does not have a "uid" attribute, ever.

Can the warning be removed without losing anything? Considerations:

* In a RFC2307-only environment, if "ldap_user_search_filter" is erroneously
  set to yield non-user objects, showing K-1540 legitimately points
  out the misconfiguration.

* RFC2307-style groups are simply unable to represent GIG, but KC
  will always attempt GIG resolution.

* RFC2256-style groups use DNs; DNs always exist. No K-1504 would be emitted
  even if there is an erroneous member in a group.

Conclusion: With RFC2256 groups, you never get a warning even if
something is broken, and with RFC2307 you always get a warning even
if nothing is broken.

The warning is hereby suppressed.

References: KC-1589