common: avoid fchmod unspecified behavior
POSIX says that fchmod on sockets is unspecified beahavior.
http://pubs.opengroup.org/onlinepubs/9699919799/functions/fchmod.html

On some implementations, mode flags will not be applied if called on
a socket's fd, resulting in local sockets that stay inaccessible.
This is a regression introduced with commit
kopanocore-8.7.82-97-gf960651e2 . This patch reverts to previous
behaviour and calls chmod on the socket's sun_path (if available);
this seems to be an acceptable fix.

References: GH-PR-19
Signed-off-by: Armin Schöffmann <armin.schoeffmann@aegaeon.de>