client: implement more automatic deletion of KCmdProxy
First, KCmdProxy is subclassed into KCmdProxy2, and the virtual
destructor helps automatically disposing of soap_endpoint.

The custom subclass also allows for marking as "final". We could not
do that with KCmdProxy (soapKCmdProxy.h), as it is autogenerated and
the change would be overwritten. Devirtualization can only be applied
when the object is used through a KCmdProxy2 pointer, so the code is
switched to use that.