common: fix ECCache reporting its own size too big
ECCache's size calculation adds the container's own size for every
element it contains, which is wrong. A pathalogical case:

	#include <map>
	#include <memory>
	#include <ECCache.h>
	struct elem : public KC::ECsCacheEntry {};
	struct map : public std::map<int, elem> { char buf[1000000]; };
	int main()
	{
		auto c = std::make_unique<KC::ECCache<map>>("test", 256000, 300);
		c->AddCacheItem(0, {}); printf("%zu\n", c->Size());
		c->AddCacheItem(1, {}); printf("%zu\n", c->Size());
		c->AddCacheItem(2, {}); printf("%zu\n", c->Size());
	}

This reports: 1000064 2000128 3000192, which is outrageous.

This could lead to underutilization, as a cache would report itself
full earlier when it is not.

valgrind says (some libc overhead included too):
"total heap usage: 48 allocs, 48 frees, 1,088,825 bytes allocated".
So it definitely cannot be 3 MB.

More sensible values are now output: 1000064 1000080 1000096.

References: KC-1531