server: add internal attachment_storage=auto
ECConfigImpl does not distinguish between defaulted and set values.
A magic value can be used however to determine un-overriden config
values, as is the case with "default:" for the KC client portion
for example.
This strategy is now used for the "attachment_storage" key to
determine when the backend is set explicitly.

References: KC-567