pyko: add text and onlineMeetingUrl setter to Ocurrence(s)
Grapi uses both appointments which is a mixin of an item and ocurrences
for calendar view endpoints. Since an occurrence is not an mixin of an
item and has it's own setters for values which are overriden in an
occurence and need to be handled different. General properties such as
the onlineMeetingUrl and text can be get from an occurence and show the
value of the 'parent' meeting but can't be set since it's not a mixin.
By adding property/setters in the Occurrence class it's now possible for
the calendar to set these properties. This solution is suboptimal and
should be improved in the future when it's required to set more
properties in the calendar on an Ocurrence.