pyko: Fix get of public:OnlineMeetingExternalLink property
I noticed that pyko was always returning NotFound for the property
public:OnlineMeetingExternalLink even if said property was set on
a given item.

It turned out that the property getter for that property did not
specify the proptype parameter, and thus did not map the property
to the value it was saved to by the setter.

This change cleans and unifies getter and setter parameters which
fixes the issue.