php-text: include basic unittest using run-tests.php
PHP C extensions can be tested using run-tests.php which is normally
available when using phpize and tooling and is now vendored in the repo.

Basic unittests which do not require a server session has been added and
should ater be automatically tested on C++ changes using the Jenkinsfile
logic.