build: record config KC was built with
Make it impossible for external programs to build against the KC
libraries when a different C++ standard was selected.

This is because KC defines some of its type aliases depending on the
chosen C++ standard, which means there are places where the ABI
changes, e.g. sizeof(ECLogger_File).

Furthermore, include "config.h" must not appear in exported headers,
because config.h is a namespace polluter and that include statement
is prone to pick up another project's config.h when the search
directories are right. The fix is to establish a second config header
that is well-controlled (<kopano/buildconfig.h> here).