client: expose sync log messages to real world
The change exporter has had log messages for its entire life, but
this was only ever enabled on user demand with the
ZARAFA_SYNC_LOGLEVEL environment variable, and that got disabled with
refactoring in kopanocore-8.6.81-85-g34ffc6c29 .

Make this available again, and, rather than emit this to a file,
integrate it into the regular logger so that it goes to the same
stream as the regular log, which has the benefit that one does not
have to piece together two different log files. A new environment
variable (KOPANO_CLIENT_LOGLEVEL) and logging component bit is added
for this purpose.

Fixes: kopanocore-8.6.81-85-g34ffc6c29
References: KF-2269