dagent: add "single" process model for debugging
Since python is not thread-safe, it is generally incompatible with
thread process mode. However, the fork process model is undesirable
too in debugging, as gdb does not support self-cloning like perldebug
when the inferior generates a child.

One way to get around this conundrum is to just run everything
serially without any parallelization whatsoever.