client: avoid eager property retrieval when asking for proplist
Problems like KC-1652 and KC-1653 (or, for example, a different issue
in librosie) can show not only when a program explicitly requests
PR_EC_BODY_FILTERED, but also when merely asking for the list of
available properties on a message.

This is because GetPropList goes to compute all the property values.
A source code comment already points out that this is wasteful and
not strictly necessary.

Doing away with that yields a large improvement of 43% as measured on
a 3700X CPU. The client-side C API alone (cached; no network traffic)
reaches almost +1500%. This is with the default config where tidied
HTML is only produced on-demand and on-the-fly at all times.

\                        proplist     inv    proplist1     inv
Full computation              640ops 1562µs       6570ops  152µs
Remove EC_BODY_FILTERED       840    1190µs      46300      22µs
Remove pre-computation        920    1087µs      98500      10µs

Improvement                    43%                1499%

References: KC-1653, KC-1660