daemons: add missing SIGHUP handling parts
When a daemon runs in forked mode, SIGHUP needs to be handled by both
the main process, plus every worker (due to the distinct memory space
and file descriptor table), so g_sighup_flag checks are needed in
both the connection dispatcher loop and the (per-client) command wait
loop.

This also creates a new issue: in threaded mode, g_sighup_sync
becomes susceptible to multiple execution, which needs to be fenced
off with a one-off variable.

The execution need not be restrited to the main thread though.

Fixes: kopanocore-8.7.9-58-gf31a887af
References: KC-1654