storeadm: drop pointless display_name_w in output
For every mailbox with a name, storeadm would emit

	"display_name_w": true

This is because jsoncpp has no explicit conversion routine for
wchar_t *, so it was downconverted to bool. This makes the property
completely useless and is hereby removed.

Because UTF-8 is backwards compatible and JSON is UTF-8, the
display_name property is upgraded to the Unicode name.