fsck: consistent unique entity check
By trivially concatenating the display name with the email address,
two individual entities

	Foo <bar@baz.de>
	Fooba <r@baz.de>

would be considered the same, even though they are not. This can be
reproduced by creating an appointment, adding the two attendees to
the invite list and then saving the appointment (without sending the
invitation message).

References: KC-1665