common: check whether ECChannel SSL context could be created
Currently, a new SSL context is initialized with the corresponding
callbacks, but the returned pointer from SSL_CTX_new is never
actually checked for a success return value, so it might cause a
SIGSEGV later on. Add a check which tests the return value of
SSL_CTX_new().

References: KF-2886