Remove useless SSL_* calls and follow documentation for SSL_accept()
A newly created SSL instance does not need to be cleared before use (remove
SSL_clear()), and does not need to be set to accept state (SSL_accept() does
this, too). The documentation for SSL_accept() explicitly states that
SSL_shutdown() must not be called on an invalid SSL instance, except when
SSL_accept() ran and had a return code different from the two being checked.