Implement new-style SSL/TLS version setup for OpenSSL 1.1.x+
OpenSSL 1.1.x+ comes with a new API to select the SSL/TLS versions that
are available for a TLS/SSL server. This patch implements an Apache-style
heuristic to select the corresponding min/max versions that are selected
from a protocol inclusion/exclusion list.

Additionally, the patch disables SSLv2 completely, and also adds
SSL_NO_COMPRESSION to the default set of parameters set up for the SSL
context.