pyko: recurrence: declare before parsing
When an 'invalid' recurrence is parsed we catch the exception and set
parsed to false. This however does not stop item.occurrences() from
getting called and expecting self._exceptions to be declared.

This commit is a workaround, a more proper fix would be declaring all
variables in the __init__ function before parsing.

References: KC-1206