pyko: server: handle https port not being 237
The server code is very naive in that it hardcodes the only valid https
port to be 237 which can be freely configured by the system
administrator. Rely on urllib to parse the server_socket and obtain the
valid https port when found otherwise fall back to 237. The code lacks
raising an exception when the serverport is not 237 but a file://
socket.