libserver: disable PR_ACCESS shortcut for GetContentsTable entries
The PR_ACCESS as presented by a message table (e.g. GetContentsTable
for a folder), ZCP/KC takes a shortcut by not reading the actual ACL
for the message(s), but synthesizes a summary value for PR_ACCESS.

(This is not security relevant — permissions are still checked when
the message is actually opened.)

The following testcase:

	a. Open public folder
	b. Determine rights for message M1 (possibly synthesized)
	c. Determine rights for message M2 (possibly synthesized)
	d. Open message M1
	e. Open message M2

produces the following values (not measured to statistically satisfying levels / not enough runs,
but do provide a rough idea):

               COLD READ               HOT READ
               a   b   c   d  e  sum   a   b   c   d   e  sum
With Shortcut  7- 17- 10-134-18  186   7- 11-  6- 10- 11   45
No Shortcut    7-136-114- 14-13  284   3- 15-  9- 14- 11   52
               (all microseconds)

So there is a ~50% time penalty for cold cases.

References: KC-1755