libserver: always reject tproperty request for PR_HTML
When PR_HTML/PR_RTF_COMPRESSED is requested for a table view, it is
only served when it is already in the cell cache, that is, if any
client has read the property earlier by different means
(GetProps/OpenProperty), and is never served from tproperties, because
the table does not have it by definition.

Change this to always return an error, similar to Exchange, because
cache-dependent behavior is really uncanny.

References: KS-43226, KS-43977, KS-44516, KC-1765