treewide: avoid haphazard variants of ADL
Even though no "using std::find" is in effect, "find" can get
resolved to std::find due to ADL, but (in this case) it is fragile,
because it is keying off the namespace of the return types of the
begin() and end() functions. If those return types happen to be
aliases for e.g. int *, ADL would fail.

Qualify these algorithm calls with std::.