mapi4linux: remove cross-open store cache
The keeping of a map store in M4LMAPISession causes a reference cycle, as all
IMsgStores are children of the session. (The problem was already known around
SVN r10187 (Apr 24 2008), though did not manifest in the past due to the broken
guarantee of IMAPISession going away.)

Break the cycle for real this time. There are only two users of
IMAPISupport::OpenEntry. Caching can be reimplemented in zcontacts, and the
other place in ECMsgStore seems like a corner case (setting read flag of a
draft message) that does not need to be fast.

References: KC-1791