caldav: shrink object code through CSE
By rewriting constructs like

	if (a) func(e1, e2); else func(e1, e3);

to

	func(e1, a ? e2 : e3);

it is possible to save a lot of generated object code. (Compilers do
not seem to be able to do this kind of subexpression eliminiation on
their own presently).

By chance I found https://stackoverflow.com/a/59330587 , and when an
assignment+return can be replaced by return (and lose bytes), why
not.

» size .libs/kopano-ical
   text    data     bss     dec     hex filename
 306299    5040     584  311923   4c273 .libs/kopano-ical before
 302798    5048     584  308430   4b4ce .libs/kopano-ical after
  -3501      +8      ±0   -3493         change