common: stop limiting db rollback to success scopes
The uncaught_exceptions() check was once added because, if there is a
std::bad_alloc exception, the rollback is presumed to also end up
throwing std::bad_alloc for the same reason (lack of memory).

But: The occurrence of a certain exception does not mean that the
rollback will also fail. Consider that std::logic_error occurs; then,
rollback is likely going to execute successfully.