m4lcommon: avoid losing rows from ECMemTable
move(iterRows->second.lpsPropVal), as it indicates, detached the old
props from the row. But if HrCopyPropertyArray runs into an error,
the old props are not attached back into the row.

Resolve that by deferring the prop switch until after the copy is
known to have completed successfully.

Fixes: kopanocore-8.7.81~158^2~1
References: KS-43448