mapi4linux: drop DUMB_MAPIALLOC
[My attempt at best explanation of how all this offsetof mess
went.]

offsetof/__builtin_offsetof throws a warning when used with a C++
class that is not StandardLayout, because for that case, offsetof is
undefined behavior.

It is said that offsetof could not be implemented using standard C++
constructs alone.

  * Reportedly, to know &x.m (for an expression like "(char *)&x.m -
    (char *)x", a concrete x is needed, which does not exist yet at
    compile-time.

  * The offsetof API does not even ask for a concrete object though,
    but, in fact, a type...

  * An expression like the common C implementation of offsetof,
    "(size_t)&((X *)0)->m", does not specify a valid X instance at
    any time.

As a result, the guts of offsetof are compiler-dependent.

GCC has

	#define offsetof(t, m) __builtin_offsetof(t, m)

and __builtin_offset warns if being used with a non-StandardLayout
class in accordance with the C++ specification.

SVN r1954 (2005-10-26) introduced a custom macro offsetof1 to
circumvent the warning

	#define offsetof1(t, m) ((size_t)(&((t *)1)->m)-1)

This runs into the bullet point problems mentioned above.
Subsequently, clang and Coverity warned about the use of offsetof1
__even for StandardLayout classes__, such as the MAPI allocator's
linkmap (introduced in bp/8.3.0-4-ge08601cdb). The remedy at the time
was to introduce an allocator with less intricate layout just for the
static analyzer (bp/8.3.0-256-g1ec37cbf0).

In retrospect, METHOD_PROLOGUE would have just needed to use a
offsetof1-based container_of, while the MAPI allocator use a
real-offsetof-based container_of to suppress all warnings.

METHOD_PROLOGUE was ultimately removed (in
kopanocore-10.0.5-89-gba813ecb7), and container_of switched back to
real offsetof, which lead to the analyzer warnings disappearing, as
outlined above. As a result, DUMB_MAPIALLOC is also no longer needed
now.