libserver: do not call waitpid on negative PID
If ntlm_auth did not even start in the first place, m_NTLM_pid may
carry the errno code, and that value may be any negative number, in
which case waitpid should not be called.

This fix is of cosmetic nature only, resolving the emission
of two nonsensical log messages.

	"""
	Removing ntlm_auth on pid -12. Exitstatus: 1
	ntlm_auth was killed by signal 1
	"""

References: KC-45863