gateway: fix null dereference when IMAP properties are incomplete
bSkipOpen starts out as true. If a datum is requested that is not in
lpProps, it needs to be synthesized on-the-fly from the message,
setting bSkipOpen=false.

But bSkipOpen is prone to getting erroneously set back to true:
Suppose that the message has no PR_EC_IMAP_BODY but does have
PR_EC_IMAP_EMAIL_SIZE. Given lstDataItems = {BODY, RFC822.SIZE},
there will be a crash when BODY is requested due to bSkipOpen=true
and therefore lpMessage=nullptr.

CID-199574 (#1 of 1): Explicit null dereferenced (FORWARD_NULL)
82. var_deref_model: Passing null pointer lpMessage to OpenProperty,
which dereferences it. (The dereference happens because this is a
virtual function call.)