server: fix kustomer wait timeout check
The comparison for the kustomer_initialize_timeout configuration setting
uses a invalid comparison. It does work since the used value is
equivalent to the 0 value which the comparison should use. This change
corrects the copy/paste error and does replace the comparison value with
a real 0 to avoid possible confusion.

Reference: KC-1812